+++ modules/@apostrophecms/page/ui/apos/components/AposPagesManager.vue
@@ -21,6 +21,18 @@
         @click="confirmAndCancel"
       />
     </template>
+    <template
+      v-if="showLocalePicker"
+      #localeDisplay
+    >
+      <AposDocLocalePicker
+        :locale="modalData.locale"
+        :module-options="moduleOptions"
+        :is-modified="false"
+        :show-localized="false"
+        @switch-locale="switchLocale"
+      />
+    </template>
     <template #primaryControls>
       <AposUtilityOperations
         :module-options="moduleOptions"